+++ src/camera.rs
@@ -5,23 +5,23 @@ use num::traits::Zero;
 #[derive(Debug, Clone, Copy)]
 pub struct Camera {
     /// Camera eye position
-    pub eye: Vec3<f64>,
+    pub eye: Vec3<f32>,
     /// Camera view direction vector
-    pub dir: Vec3<f64>,
-    pub up: Vec3<f64>,
-    pub fov: f64,
+    pub dir: Vec3<f32>,
+    pub up: Vec3<f32>,
+    pub fov: f32,
     pub width: u32,
     pub height: u32,
 }
 
 pub struct CamBuilder {
     /// Camera eye position
-    eye: Vec3<f64>,
+    eye: Vec3<f32>,
     /// Point to which camera looks at
-    center: Vec3<f64>,
-    up: Vec3<f64>,
+    center: Vec3<f32>,
+    up: Vec3<f32>,
     /// View direction, i.e., center - eye vector
-    fov: f64,
+    fov: f32,
     width: u32,
     height: u32,
 }
@@ -39,22 +39,22 @@ impl CamBuilder {
         }
     }
 
-    pub fn eye(&mut self, eye: Vec3<f64>) -> &mut CamBuilder {
+    pub fn eye(&mut self, eye: Vec3<f32>) -> &mut CamBuilder {
         self.eye = eye;
         self
     }
 
-    pub fn center(&mut self, center: Vec3<f64>) -> &mut CamBuilder {
+    pub fn center(&mut self, center: Vec3<f32>) -> &mut CamBuilder {
         self.center = center;
         self
     }
 
-    pub fn up(&mut self, up: Vec3<f64>) -> &mut CamBuilder {
+    pub fn up(&mut self, up: Vec3<f32>) -> &mut CamBuilder {
         self.up = up;
         self
     }
 
-    pub fn fov(&mut self, fov: f64) -> &mut CamBuilder {
+    pub fn fov(&mut self, fov: f32) -> &mut CamBuilder {
         self.fov = fov;
         self
     }
@@ -83,29 +83,29 @@ impl CamBuilder {
 }
 
 impl Camera {
-    pub fn roll(&mut self, angle: f64) {
+    pub fn roll(&mut self, angle: f32) {
         let rot = Rot3::from_axis_angle(&Unit::new_normalize(self.dir), angle.to_radians());
         self.up = (rot * self.up).normalize();
     }
 
-    pub fn yaw(&mut self, angle: f64) {
+    pub fn yaw(&mut self, angle: f32) {
         let rot = Rot3::from_axis_angle(&Unit::new_normalize(self.up), angle.to_radians());
         self.dir = rot * self.dir;
     }
 
-    pub fn pitch(&mut self, angle: f64) {
+    pub fn pitch(&mut self, angle: f32) {
         let side = self.up.cross(&self.dir);
         let rot = Rot3::from_axis_angle(&Unit::new_normalize(side), angle.to_radians());
         self.dir = rot * self.dir;
         self.up = self.dir.cross(&side).normalize();
     }
 
-    pub fn mov_fwd(&mut self, dist: f64) {
+    pub fn mov_fwd(&mut self, dist: f32) {
         let dir = self.dir;
         self.eye = self.eye + dir * dist;
     }
 
-    pub fn mov_side(&mut self, dist: f64) {
+    pub fn mov_side(&mut self, dist: f32) {
         let dir = self.dir;
         let side = self.up.cross(&dir);
         self.eye = self.eye + side * dist;
